app/requirements.txt
asgiref==3.4.1
boto3==1.20.33
cassandra-driver==3.25.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cors==1.0.1
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0
django-cassandra-engine==1.6.3
django-cors-headers==3.10.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
django-storages==1.12.3
djoser==2.1.0
filelock==3.4.0
future==0.18.2
geomet==0.2.1.post1
gevent==21.12.0
greenlet==1.1.2
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
minio==6.0.2
oauthlib==3.1.1
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tldextract==3.1.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
zope.event==4.5.0
zope.interface==5.4.0
mysqlclient

